% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{constitucion_bbdd}
\alias{constitucion_bbdd}
\title{Conexion a la Base de Datos de Constitucion Abierta}
\usage{
constitucion_bbdd(dir = constitucion_path())
}
\arguments{
\item{dir}{La ubicacion de la base de datos en el disco. Por defecto es
\code{constitucionabierta} en la carpeta de datos del usuario (\code{\link[rappdirs:user_data_dir]{rappdirs::user_data_dir()}}),
o la variable de entorno \code{CONSTITUCION_BBDD_DIR}.}
}
\description{
Devuelve una conexion a la base de datos local. Esto corresponde a una
conexion a una base 'SQLite' compatible con DBI. A diferencia de
\code{\link[=constitucion_tabla]{constitucion_tabla()}}, esta funcion es mas flexible y se puede usar con
dbplyr para leer unicamente lo que se necesita o directamente con DBI para
usar comandos SQL.
}
\examples{
if (constitucion_estado()) {
 DBI::dbListTables(constitucion_bbdd())

 DBI::dbGetQuery(
  constitucion_bbdd(),
  'SELECT * FROM comunas WHERE provincia_ref_id = 1'
 )
}
}
