% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{constitucion_borrar_base}
\alias{constitucion_borrar_base}
\title{Elimina la Base de Datos de Constitucion Abierta de tu Computador}
\usage{
constitucion_borrar_base()
}
\description{
Elimina el directorio \code{constitucionabierta} en el directorio de datos de usuario.
}
\examples{
\donttest{
\dontrun{
constitucion_borrar_base()
}
}
}
