% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{constitucion_descargar_base}
\alias{constitucion_descargar_base}
\title{Descarga la Base de Datos de Constitucion Abierta a tu Computador}
\usage{
constitucion_descargar_base(ver = NULL, borrar = TRUE)
}
\arguments{
\item{ver}{La version a descargar. Por defecto es la ultima version disponible en GitHub.
Se pueden ver todas las versiones en
\url{https://github.com/pachamaltese/constitucionabierta/releases}.}

\item{borrar}{Borrar o no el archivo zip luego de cargar la base de datos.}
}
\description{
Este comando descarga la base de datos completa como un unico archivo bz2 que
se descomprime para dejar disponible la base de datos local. La descarga son 15 MB
y la base de datos usa 65 MB en disco.
}
\examples{
\donttest{
\dontrun{
constitucion_descargar_base()
}
}
}
