% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{constitucion_desconectar_base}
\alias{constitucion_desconectar_base}
\title{Desconecta la Base de Datos de Constitucion Abierta}
\usage{
constitucion_desconectar_base()
}
\description{
Una funcion auxiliar para desconectarse de la base de datos.
}
\examples{
constitucion_desconectar_base()
}
