% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-pane.R
\name{constitucion_panel}
\alias{constitucion_panel}
\title{Abre el Panel de Conexion a la Base de Datos de Constitucion Abierta en RStudio}
\usage{
constitucion_panel()
}
\description{
Esta funcion abre el panel "Connections" para explorar la base de datos
local de forma interactiva.
}
\examples{
if (!is.null(getOption("connectionObserver"))) constitucion_panel()
}
